userspace: u64 signal masks (RT signals 33-63) + 9 new Linux syscalls
Signal bitmaps widen u32 -> u64 (SIGNAL_PENDING / SIGNAL_MASK /
SIGRETURN_SAVED_MASK, NSIG 32 -> 64): RT signals are now sendable
(kill/tkill/tgkill/sigqueue rejected everything >= 32), blockable
(sigprocmask/sigsuspend/ppoll/epoll_pwait truncated the sigset to 32
bits — musl's SIGCANCEL=33 pthread_cancel block silently did nothing),
handleable (64-slot sigaction table), and reportable
(rt_sigpending/rt_sigtimedwait/signalfd). Bit-N-=-signal-N caps the
range at 63; signal 64 (SIGRTMAX) stays EINVAL, documented at
SIGNAL_PENDING. set_signal_mask_for_task now strips SIGKILL/SIGSTOP on
every install path, not just sigprocmask's.

New syscalls: gettimeofday(96/169), settimeofday(164/170), time(201),
inotify_init(253, must NOT read the flags register — the legacy ABI
has no args), ioprio_set/get(251-252, 30-31, store-and-return),
fchdir(81/50), link(86) / linkat(265/37) over the new DirOps::link
(same-parent like rename; cross-directory -> -EXDEV).

/proc/[pid]/exe published from execve's post-shebang cur_path
(PROC_EXE table, swept at exit); exe/cwd/root hooks wired in
cross_crate_init.

Harness fixes surfaced by the suite's hash-order reshuffle:
own_stack_wait_child's no-executor degrade arm now clears
wait_child_pending (a stale flag on the persistent FAKE_TASK uctx
misrouted a later pause(2) into the wait4 path); pause_neg accepts
both -EINTR (yield hook installed) and -1 (no-executor fallback) —
which one appears is unrelated-test order, not behavior; two stale
NSIG=32 boundary asserts updated (signal 32/40 are valid now).

kernel-test: 5973 pass / 0 fail / 65 skip; clippy -D warnings + fmt
clean across the CI feature matrix.

+++ userspace/src/epoll.rs
@@ -640,7 +640,7 @@ fn epoll_wait_common(ctx: &mut dyn TrapContext, is_pwait: bool) {
         // SAFETY: sigsetsize == 8 checked above; copy_from_user range-validates
         // `sigmask_ptr` and SMAP-brackets the read into `buf`.
         if unsafe { crate::handlers::copy_from_user(&mut buf, sigmask_ptr) }.is_ok() {
-            let mask = (u64::from_ne_bytes(buf) << 1) as u32;
+            let mask = u64::from_ne_bytes(buf) << 1;
             old_mask = Some(crate::handlers::set_signal_mask_for_task(task, mask));
         } else {
             ctx.set_return(fail);
